+++ lstchain/mc/tests/test_senstivity.py
@@ -20,7 +20,6 @@ from lstchain.mc.sensitivity import (
     find_cut_real
 )
 
-
 @pytest.mark.run(after='test_r0_to_dl1')
 def test_read_sim_par(simulated_dl1_file):
     par = read_sim_par(simulated_dl1_file)
@@ -124,8 +123,42 @@ def test_ring_containment():
     np.testing.assert_equal(contained, [False, True, True, True, True,
                                         False, False, False, False, False])
 
-@pytest.mark.run(after='test_apply_models')
-def test_sens():
-    # TODO: define test for sens
-    # dl2 = pd.read_hdf(dl2_file, key=dl2_params_lstcam_key)
-    pass
+
+@pytest.fixture(scope="session")
+def test_sensitivity(fake_dl2_proton_file, simulated_dl2_file):
+    from lstchain.mc.sensitivity import (sensitivity_gamma_efficiency,
+                                         sensitivity_gamma_efficiency_real_protons,
+                                         sensitivity_gamma_efficiency_real_data)
+    
+    geff_gammaness = 0.9
+    geff_theta2 = 0.8
+    eb = 10
+    obstime = 50 * 3600 * u.s
+    noff = 2
+    
+    sensitivity_gamma_efficiency(simulated_dl2_file,
+                                 fake_dl2_proton_file,
+                                 1, 1,
+                                 eb,
+                                 geff_gammaness,
+                                 geff_theta2,
+                                 noff,
+                                 obstime)
+    
+    sensitivity_gamma_efficiency_real_protons(simulated_dl2_file,
+                                              fake_dl2_proton_file,
+                                              1,
+                                              eb,
+                                              geff_gammaness,
+                                              geff_theta2,
+                                              noff,
+                                              obstime)
+    
+    sensitivity_gamma_efficiency_real_data(simulated_dl2_file,
+                                           fake_dl2_proton_file,
+                                           0.5,0.5,
+                                           eb,
+                                           geff_gammaness,
+                                           geff_theta2,
+                                           noff,
+                                           obstime)
